mysite/pyproject.toml
[tool.poetry]
name = "blogapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.7"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
python-dotenv = "1.0.1"
asgiref = "3.8.1"
pyjwt = "2.8.0"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
tzdata = "2024.1"
stripe = "^10.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
